// apps/admin/src/shared/api/generated/logs/schemas/exportFileDownloadsParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — Logs subset
 * Filtered for Homepage Admin logs Orval codegen.
 * OpenAPI spec version: v1
 */
import type { ExportFileDownloadsAssetPurpose } from './exportFileDownloadsAssetPurpose';

export type ExportFileDownloadsParams = {
adminName?: string;
fileName?: string;
assetPurpose?: ExportFileDownloadsAssetPurpose;
sourceType?: string;
from?: string;
to?: string;
};