// apps/admin/src/shared/api/generated/logs/schemas/fileDownloadsParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — Logs subset
 * Filtered for Homepage Admin logs Orval codegen.
 * OpenAPI spec version: v1
 */
import type { FileDownloadsAssetPurpose } from './fileDownloadsAssetPurpose';

export type FileDownloadsParams = {
adminName?: string;
fileName?: string;
assetPurpose?: FileDownloadsAssetPurpose;
sourceType?: string;
from?: string;
to?: string;
page?: number;
size?: number;
};